test: keep integration-tests package
