+++ packages/integration-tests/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@edge-runtime/integration-tests",
+  "private": true,
+  "version": "1.0.0",
+  "scripts": {
+    "test": "pnpm run test:node && pnpm run test:edge",
+    "test:edge": "jest --testEnvironment @edge-runtime/jest-environment",
+    "test:node": "jest --testEnvironment node"
+  },
+  "license": "MPL-2.0",
+  "devDependencies": {
+    "@edge-runtime/jest-environment": "workspace:*",
+    "@edge-runtime/ponyfill": "workspace:*",
+    "@edge-runtime/primitives": "workspace:*"
+  }
+}
