+++ packages/integration-tests/test/blob.test.ts
@@ -3,6 +3,6 @@ const test18 = process.version.startsWith('v18.') ? test : test.skip
 test18('Blob is available globally after importing the ponyfill', () => {
   const blob = Blob
   expect(blob).toBeDefined()
-  require('..')
+  require('@edge-runtime/ponyfill')
   expect(Blob).toBe(blob)
 })
