+++ packages/integration-tests/test/headers.test.ts
@@ -1,6 +1,4 @@
-/**
- * @jest-environment ../jest-environment/dist
- */
+import { Headers } from '@edge-runtime/ponyfill'
 
 test('sets header calling Headers constructor', async () => {
   const headers = new Headers({ cookie: 'hello=world' })
