+++ packages/integration-tests/test/url.test.ts
@@ -1,6 +1,4 @@
-/**
- * @jest-environment ../jest-environment/dist
- */
+import { URL } from '@edge-runtime/ponyfill'
 
 test('URL', async () => {
   const url = new URL('https://edge-ping.vercel.app/')
