+++ packages/vm/tests/integration/crypto.test.ts
@@ -1,22 +1,4 @@
 import { EdgeVM } from '../../src'
-import { createHash } from 'crypto'
-
-function toHex(buffer: ArrayBuffer) {
-  return Array.from(new Uint8Array(buffer))
-    .map((b) => b.toString(16).padStart(2, '0'))
-    .join('')
-}
-
-test('crypto.randomUUID', async () => {
-  const vm = new EdgeVM()
-
-  async function fn() {
-    return crypto.randomUUID()
-  }
-
-  const uuid = await vm.evaluate(`(${fn})()`)
-  expect(uuid).toEqual(expect.stringMatching(/^[a-f0-9-]+$/))
-})
 
 test('crypto.subtle.digest returns an ArrayBuffer', async () => {
   const vm = new EdgeVM()
@@ -35,23 +17,6 @@ test('crypto.subtle.digest returns an ArrayBuffer', async () => {
   expect(digest).toBeInstanceOf(fromContext.ArrayBuffer)
 })
 
-test('crypto.subtle.digest returns a SHA-256 hash', async () => {
-  const vm = new EdgeVM()
-
-  async function fn() {
-    const digest = await crypto.subtle.digest(
-      'SHA-256',
-      new Uint8Array([104, 105, 33]),
-    )
-    return digest
-  }
-
-  const digest = await vm.evaluate(`(${fn})()`)
-  expect(toHex(digest)).toEqual(
-    createHash('sha256').update('hi!').digest('hex'),
-  )
-})
-
 test('crypto.generateKey works with a Uint8Array from the VM', async () => {
   async function fn() {
     await crypto.subtle.generateKey(
@@ -68,31 +33,4 @@ test('crypto.generateKey works with a Uint8Array from the VM', async () => {
 
   const vm = new EdgeVM()
   await vm.evaluate(`(${fn})()`)
-})
-
-test('Ed25519', async () => {
-  const vm = new EdgeVM()
-
-  function fn() {
-    return crypto.subtle.generateKey('Ed25519', false, ['sign', 'verify'])
-  }
-
-  const kp = await vm.evaluate(`(${fn})()`)
-  expect(kp).toHaveProperty('privateKey')
-  expect(kp).toHaveProperty('publicKey')
-})
-
-test('X25519', async () => {
-  const vm = new EdgeVM()
-
-  function fn() {
-    return crypto.subtle.generateKey('X25519', false, [
-      'deriveBits',
-      'deriveKey',
-    ])
-  }
-
-  const kp = await vm.evaluate(`(${fn})()`)
-  expect(kp).toHaveProperty('privateKey')
-  expect(kp).toHaveProperty('publicKey')
 })
